pyproject.toml
[project]
name = "resume-screener-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.4",
    "google-generativeai>=0.8.5",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "streamlit>=1.52.1",
    "uvicorn>=0.38.0",
]
